Create own file for markdown generation function
